pyproject.toml
[project]
name = "codepicture"
version = "0.1.0"
description = "Generate beautiful images of source code"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.5",
    "pygments>=2.19",
    "catppuccin[pygments]>=2.5",
    "pycairo>=1.25",
    "manimpango>=0.6.1",
    "pillow>=10.0",
    "typer>=0.21.1",
]

[project.scripts]
codepicture = "codepicture.cli:app"

[project.entry-points."pygments.lexers"]
mlir = "codepicture.highlight.custom_lexers.mlir_lexer:MlirLexer"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codepicture"]

[tool.hatch.build.targets.wheel.force-include]
"src/codepicture/fonts/JetBrainsMono-Regular.ttf" = "codepicture/fonts/JetBrainsMono-Regular.ttf"

[dependency-groups]
dev = [
    "cairosvg>=2.8.2",
    "pixelmatch>=0.3.0",
    "pre-commit>=4.0",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.0.0",
    "pytest-randomly>=4.0.1",
    "pytest-timeout>=2.4.0",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --benchmark-skip"
timeout = 5
timeout_method = "signal"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/codepicture"]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_also = [
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "class .*\\bProtocol\\):",
]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]
# Visual fixtures are rendered as images; any change alters baselines
extend-exclude = ["tests/visual/fixtures"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
# Pygments lexer class attributes follow the Pygments API convention
"src/codepicture/highlight/custom_lexers/mlir_lexer.py" = ["RUF012"]
# Benchmark fixtures are intentional code samples, not production code
"tests/benchmarks/fixtures/*.py" = ["B007", "B904", "F841", "SIM", "UP046", "RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["codepicture"]

[tool.ruff.format]
quote-style = "double"
